Factor common udevadm interactions into utils subdirectory.
