+++ tests/plugins/udev_database.py
@@ -30,290 +30,9 @@
 from __future__ import (print_function, division, unicode_literals,
                         absolute_import)
 
-import sys
-import os
-import re
-import errno
-import random
-import subprocess
-from collections import Iterable, Sized
-
 import pytest
 
-
-class UDevAdm(object):
-    """
-    Wrap ``udevadm`` utility.
-    """
-
-    CANDIDATES = ['/sbin/udevadm', 'udevadm']
-    _adm = None
-
-    @classmethod
-    def find(cls):
-        """
-        Construct a valid :class:`UDevAdm` object.
-
-        :returns: a working :class:`UDevAdm` object
-        :rtype: :class:`UDevAdm`
-        :raises EnvironmentError:
-        """
-        for candidate in cls.CANDIDATES:
-            try:
-                udevadm = cls(candidate)
-                # try to execute udev to make sure that it's actually
-                # executable
-                udevadm.query_udev_version()
-                return udevadm
-            except EnvironmentError as error:
-                if error.errno != errno.ENOENT:
-                    raise
-
-    @classmethod
-    def adm(cls):
-        """
-        Returns the singleton object of this class, if one can be found.
-
-        :returns: singleton :class:`UDevAdm` object
-        :rtype: :class:`UDevAdm`
-        """
-        if cls._adm is None:
-            try:
-                cls._adm = cls.find()
-            except EnvironmentError:
-                pass
-        return cls._adm
-
-    def __init__(self, udevadm):
-        """
-        Create a new ``udevadm`` wrapper for the given udevadm executable.
-
-        ``udevadm`` is the path to udevadm as string.  If relative, ``udevadm``
-        is looked up in ``$PATH``.
-        """
-        self.udevadm = udevadm
-
-    def query_udev_version(self):
-        return int(self._execute('--version'))
-
-    def _execute(self, *args):
-        command = [self.udevadm] + list(args)
-        proc = subprocess.Popen(command, stdout=subprocess.PIPE)
-        output = proc.communicate()[0].strip()
-        if proc.returncode != 0:
-            raise subprocess.CalledProcessError(proc.returncode, command)
-        return output
-
-    def _execute_query(self, device_path, query_type='all'):
-        output = self._execute('info', '--root', '--path', device_path,
-                               '--query', query_type)
-        return output.decode(sys.getfilesystemencoding())
-
-    def query_devices(self):
-        database = self._execute('info', '--export-db').decode(
-            sys.getfilesystemencoding()).splitlines()
-        for line in database:
-            line = line.strip()
-            if not line:
-                continue
-            type, value = line.split(': ', 1)
-            if type == 'P':
-                yield value
-
-    def query_device_properties(self, device_path):
-        properties = {}
-        for line in self._execute_query(device_path, 'property').splitlines():
-            line = line.strip()
-            property, value = line.split('=', 1)
-            properties[property] = value
-        return properties
-
-    def query_device_attributes(self, device_path):
-        output = self._execute(
-            'info', '--attribute-walk', '--path', device_path)
-        attribute_dump = output.decode(
-            sys.getfilesystemencoding()).splitlines()
-        attributes = {}
-        for line in attribute_dump:
-            line = line.strip()
-            if line.startswith('looking at parent device'):
-                # we don't continue with attributes of parent devices, we only
-                # want the attributes of the given device
-                break
-            if line.startswith('ATTR'):
-                name, value = line.split('==', 1)
-                # remove quotation marks from attribute value
-                value = value[1:-1]
-                # remove prefix from attribute name
-                name = re.search('{(.*)}', name).group(1)
-                attributes[name] = value
-        return attributes
-
-    def query_device(self, device_path, query_type):
-        if query_type not in ('symlink', 'name'):
-            raise ValueError(query_type)
-        try:
-            return self._execute_query(device_path, query_type)
-        except subprocess.CalledProcessError:
-            return None
-
-
-class DeviceData(object):
-    """
-    Data for a single device.
-    """
-
-    def __init__(self, device_path, udevadm):
-        self.device_path = device_path
-        self._udevadm = udevadm
-
-    def __repr__(self):
-        return '{0}({1})'.format(self.__class__.__name__, self.device_path)
-
-    @property
-    def sys_path(self):
-        """
-        Get the ``sysfs`` path of the device.
-        """
-        return '/sys' + self.device_path
-
-    @property
-    def properties(self):
-        """
-        Get the device properties as mapping of property names as strings to
-        property values as strings.
-        """
-        return self._udevadm.query_device_properties(self.device_path)
-
-    @property
-    def attributes(self):
-        """
-        Get the device attributes as mapping of attributes names as strings to
-        property values as byte strings.
-
-        .. warning::
-
-           As ``udevadm`` only exports printable attributes, this list can be
-           incomplete!  Do *not* compare this dictionary directly to a
-           attributes dictionary received through pyudev!
-        """
-        return self._udevadm.query_device_attributes(self.device_path)
-
-    @property
-    def tags(self):
-        """
-        Get the device tags as list of strings.
-        """
-        tags = self.properties.get('TAGS', '').split(':')
-        return [t for t in tags if t]
-
-    @property
-    def device_node(self):
-        """
-        Get the device node path as string, or ``None`` if the device has no
-        device node.
-        """
-        return self._udevadm.query_device(self.device_path, 'name')
-
-    @property
-    def device_links(self):
-        """
-        Get the device links as list of strings.
-        """
-        links = self._udevadm.query_device(self.device_path, 'symlink')
-        return links.split() if links else []
-
-    @property
-    def device_number(self):
-        """
-        Get the device number as integer or 0 if the device has no device
-        number.
-        """
-        if self.device_node:
-            return os.stat(self.device_node).st_rdev
-        else:
-            return 0
-
-
-class DeviceDatabase(Iterable, Sized):
-    """
-    The udev device database.
-
-    Takes a snapshot of the udev device database when it is initialized.
-    Consequently, some :class:`DeviceData` objects in the database may
-    not exist when the database is iterated over.
-
-    This class is an iterable over :class:`DeviceData` objects that contain the
-    data associated with each device stored in the udev database.
-    """
-
-    _db = None
-
-    @classmethod
-    def db(cls, renew=False):
-        """
-        Get a database object.
-
-        :param bool renew: if renew is True, get a new object
-        :returns: a database object
-        :rtype: :class:`DeviceDatabase`
-    """
-        if cls._db is None or renew:
-            udevadm = UDevAdm.adm()
-            if udevadm:
-                cls._db = DeviceDatabase(udevadm)
-        return cls._db
-
-
-    def __init__(self, udevadm):
-        self._udevadm = udevadm
-        self._devices = set(self._udevadm.query_devices())
-
-    def __iter__(self):
-        return (DeviceData(d, self._udevadm) for d in self._devices)
-
-    def __len__(self):
-        return len(self._devices)
-
-    def find_device_data(self, device_path):
-        """
-        Find device data for the device designated by ``device_path``.
-
-        ``device_path`` is a string containing the device path.
-
-        Return a :class:`DeviceData` object containing the data for the given
-        device, or ``None`` if the device was not found.
-        """
-        if device_path in self._devices:
-            return DeviceData(device_path, self._udevadm)
-        else:
-            return None
-
-def _get_device_sample(udev_database, device=None, sample_size=None):
-    """
-    Compute a sample of the udev device database.
-
-    :param udev_database: the udev database
-    :param device: the unique device to compute the data for
-    :param sample_size: the size of the sample to compute
-    :type sample_size: int or NoneType
-    :returns: a sample of udev devices
-    :rtype: list of :class:`DeviceData`
-    :raises ValueError: if the specified device does not exist
-
-    If no sample size and no device specified, returns all devices.
-    """
-    if device is not None:
-        device_data = udev_database.find_device_data(device)
-        if not device_data:
-            raise ValueError('{0} does not exist'.format(device))
-        return [device_data]
-    elif sample_size is not None:
-        actual_size = min(sample_size, len(udev_database))
-        return random.sample(list(udev_database), actual_size)
-    else:
-        return list(udev_database)
-
+from tests.utils import udev
 
 def pytest_runtest_setup(item):
     """
@@ -353,17 +72,17 @@ def pytest_configure(config):
         'markers', 'udev_version(spec): mark test to run only if the udev '
         'version matches the given version spec')
 
-    udevadm = UDevAdm.adm()
+    udevadm = udev.UDevAdm.adm()
     if udevadm:
         config.udev_version = udevadm.query_udev_version()
-        config.udev_database = DeviceDatabase.db()
+        config.udev_database = udev.DeviceDatabase.db()
 
         if config.option.all_devices:
             sample_size = None
         else:
             sample_size = config.option.device_sample_size
 
-        config.udev_device_sample = _get_device_sample(
+        config.udev_device_sample = udev.get_device_sample(
            config.udev_database,
            config.option.device,
            sample_size
